libs/mng_file/pyproject.toml
[project]
name = "mng-file"
version = "0.1.0"
description = "File command plugin for mng - read, write, and list files on agents and hosts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mng==0.1.7",
    "click-option-group>=0.5.6",
    "tabulate>=0.9.0",
]

[project.entry-points.mng]
file = "imbue.mng_file.plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["imbue"]

[tool.uv.sources]
mng = { workspace = true }

[tool.pytest.ini_options]
timeout_func_only = true
junit_family = "xunit1"
testpaths = ["."]
addopts = [
  "-n 4",
  "--dist=worksteal",
  "--cov=imbue.mng_file",
  "--cov-report=term-missing",
  "--cov-report=html",
  "--cov-report=xml",
  "--cov-fail-under=78",
  "--durations=20",
  "--timeout=10",
  "--timeout-method=signal",
  "--max-worker-restart=0",
  "-m not acceptance and not release",
  "--slow-tests-to-file",
  "--coverage-to-file",
]

[tool.coverage.run]
parallel = true
concurrency = ["multiprocessing", "thread"]
omit = [
  "*_test.py",
  "test_*.py",
  "*/tests/*",
  "*/conftest.py",
  "*/utils/testing.py",
]

[tool.inline-snapshot]
default-flags=["report"]
default-flags-tui=["report"]
default-flags-ide=["report"]
format-command="uv run ruff format --force-exclude --config pyproject.toml --stdin-filename {filename}"

[tool.pyright]
venvPath = "../.."
venv = ".venv"
pythonVersion = "3.11"
strict = ["**/*.py"]
